+++ packages/router-core/src/ssr/constants.ts
@@ -1,3 +1,2 @@
 export const GLOBAL_TSR = '$_TSR'
 export declare const GLOBAL_SEROVAL: '$R'
-export const TSR_SCRIPT_BARRIER_ID = '$tsr-stream-barrier'