+++ packages/router-cli/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @tanstack/router-cli
 
+## 1.167.38
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/router-generator@1.167.38
+
+## 1.167.37
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/router-generator@1.167.37
+
+## 1.167.36
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/router-generator@1.167.36
+
 ## 1.167.35
 
 ### Patch Changes